* feature/unity-18:
  App intro (#76)
  Fix trimmed range inaccuracy
  Modify FFmpeg trimming commandline
  Simplify button enable logic
  Fix overruning side effect
  Remove unused fields
  Lazify val initializer
  Setup exact video trimming
  Cleanup unneccessary files
  Revamp trimming activity
  Hotfix .gitignore leaking folder

+++ app/src/main/java/com/teamisland/zzazz/ui/ProjectActivity.kt
@@ -180,7 +180,7 @@ class ProjectActivity : AppCompatActivity(), IUnityPlayerLifecycleEvents {
         modelpath = intent.getStringExtra(TrimmingActivity.MODEL_PATH)
         val uri = Uri.parse(path)
         videoDuration = getDuration(this, uri)
-        frameCount = getFrameCount(this, uri)
+        frameCount = getFrameCount(this, uri).toInt()
         fps = frameCount / (videoDuration / 1000f)
 
         zoomLevel = float2DP(0.06f, resources)
